Rename Compiler to Builder
