+++ Ephect/Commands/BuildApplication.php
@@ -6,13 +6,14 @@ use Ephect\Apps\Egg\EggLib;
 use Ephect\Framework\Commands\AbstractCommand;
 use Ephect\Framework\Commands\Attributes\CommandDeclaration;
 
-#[CommandDeclaration(verb: "compile")]
-#[CommandDeclaration(desc: "Compile the application.")]
-class CompileApplication extends AbstractCommand
+#[CommandDeclaration(verb: "build")]
+#[CommandDeclaration(desc: "Build the application.")]
+class BuildApplication extends AbstractCommand
 {
     public function run(): void
     {
+
         $egg = new EggLib($this->application);
-        $egg->compile();
+        $egg->build();
     }
 }
