+++ Ephect/Framework/Web/Application.php
@@ -3,9 +3,9 @@
 namespace Ephect\Framework\Web;
 
 use Ephect\Framework\CLI\Console;
-use Ephect\Framework\Components\Compiler;
 use Ephect\Framework\Components\Component;
 use Ephect\Framework\Core\AbstractApplication;
+use Ephect\Framework\Core\Builder;
 use Ephect\Framework\Registry\CacheRegistry;
 use Ephect\Framework\Registry\ComponentRegistry;
 use Ephect\Framework\Registry\PluginRegistry;
@@ -25,7 +25,7 @@ class Application extends AbstractApplication
         $this->loadInFile();
 
         if (!ComponentRegistry::uncache()) {
-            $compiler = new Compiler;
+            $compiler = new Builder;
             $compiler->perform();
             $compiler->postPerform();
         }
